move src/core/server/execution-context into packages
