+++ packages/core/execution-context/core-execution-context-server-internal/BUILD.bazel
@@ -25,31 +25,22 @@ NPM_MODULE_EXTRA_FILES = [
   "package.json",
 ]
 
-# In this array place runtime dependencies, including other packages and NPM packages
-# which must be available for this code to run.
-#
-#  To reference other packages use:
-#    "//repo/relative/path/to/package"
-#    eg. "//packages/kbn-utils"
-#
-#  To reference a NPM package use:
-#    "@npm//name-of-package"
-#    eg. "@npm//lodash"
 RUNTIME_DEPS = [
+  "@npm//lodash",
+  "@npm//rxjs",
+  "//packages/kbn-config-schema",
 ]
 
-# In this array place dependencies necessary to build the types, which will include the
-# :npm_module_types target of other packages and packages from NPM, including @types/*
-# packages.
-#
-#  To reference the types for another package use:
-#    "//repo/relative/path/to/package:npm_module_types"
-#    eg. "//packages/kbn-utils:npm_module_types"
-#
-#  References to NPM packages work the same as RUNTIME_DEPS
 TYPES_DEPS = [
   "@npm//@types/node",
   "@npm//@types/jest",
+  "@npm//@types/lodash",
+  "@npm//rxjs",
+  "@npm//elastic-apm-node",
+  "//packages/kbn-config-schema:npm_module_types",
+  "//packages/kbn-logging:npm_module_types",
+  "//packages/core/base/core-base-server-internal:npm_module_types",
+  "//packages/core/execution-context/core-execution-context-common:npm_module_types",
 ]
 
 jsts_transpiler(
