+++ packages/core/execution-context/core-execution-context-server-internal/src/execution_context_container.ts
@@ -5,7 +5,8 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import type { KibanaExecutionContext } from '../../types';
+
+import type { KibanaExecutionContext } from '@kbn/core-execution-context-common';
 
 // Switch to the standard Baggage header. blocked by
 // https://github.com/elastic/apm-agent-nodejs/issues/2102
