+++ packages/core/execution-context/core-execution-context-server-internal/src/execution_context_service.test.ts
@@ -14,6 +14,7 @@ import {
 } from './execution_context_service';
 
 const delay = (ms: number = 100) => new Promise((resolve) => setTimeout(resolve, ms));
+
 describe('ExecutionContextService', () => {
   describe('setup', () => {
     let service: InternalExecutionContextSetup;
