+++ packages/core/execution-context/core-execution-context-server-mocks/src/index.ts
@@ -6,6 +6,4 @@
  * Side Public License, v 1.
  */
 
-export function foo() {
-  return 'hello world';
-}
+export { executionContextServiceMock } from './execution_context_service.mock';
