+++ packages/core/execution-context/core-execution-context-server/src/index.ts
@@ -6,6 +6,4 @@
  * Side Public License, v 1.
  */
 
-export function foo() {
-  return 'hello world';
-}
+export type { ExecutionContextSetup, ExecutionContextStart } from './contracts';
