--- src/core/server/execution_context/index.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-export type { KibanaExecutionContext } from '../../types';
-export { ExecutionContextService } from './execution_context_service';
-export type {
-  InternalExecutionContextSetup,
-  InternalExecutionContextStart,
-  ExecutionContextSetup,
-  ExecutionContextStart,
-  IExecutionContext,
-} from './execution_context_service';
-export type { IExecutionContextContainer } from './execution_context_container';
-export { config } from './execution_context_config';
